completed kata1 with scala
